% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accident_2013}
\alias{accident_2013}
\title{Accident data for 2013}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30202 rows and 50 columns.}
\source{
Coursera
\describe{
 \item{Columns}{Too many to list}
}
}
\usage{
accident_2013
}
\description{
Accident data for 2013
}
\examples{
\dontrun{
 accident_2013
}
}
\keyword{datasets}
