% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Filter data by state number, if state number is invalid error will trigger}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{state numbers to filter}

\item{year}{year of data to filter by}
}
\value{
a data.frame of filtered data
}
\description{
THis function will summarise a list of year data
}
\examples{
far_map_state()

}
