% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a character vector of filename to read}
}
\value{
this function returns a dataframe of the data
}
\description{
This is a simple function that will attempt to file into tbl df, if the file does not exist will return an error.
}
\examples{
fars_read(2013)

}
