% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in multiple files by year}
\usage{
fars_read_years(years)
}
\arguments{
\item{year}{The year to be in the name of the file}
}
\value{
year
}
\description{
This function will apply over a list of years and read in the data, if year is invalid error will trigger
}
\examples{
fars_read_years(c(2013, 2014))

}
