% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarse data by year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The years to summarise by}
}
\value{
a data.frame of summarised data
}
\description{
THis function will summarise a list of year data
}
\examples{
fars_summarize_years(c(2013, 2014))

}
