% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a filename based on year input}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year to be in the name of the file}
}
\value{
A character vector filename
}
\description{
This will create a character vector filename based on the year input
}
\examples{
make_filename(2013)

}
